backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
trimesh>=4.0.0
numpy>=1.24.0
Pillow>=10.0.0
openai>=1.0.0
reportlab>=4.0.0
python-dotenv